Both Node and Python backend organized
