+++ library/BackEnd/data/Allocation.json
@@ -1,62 +1,60 @@
 {
-    "allocation": [
-        {
-            "userId": 1,
-            "bookId": 1,
-            "userName": "Pradeep",
-            "num_copies": 2
-        },
-        {
-            "userId": 2,
-            "bookId": 1,
-            "userName": "Satish",
-            "num_copies": 1
-        },
-        {
-            "userId": 3,
-            "bookId": 2,
-            "userName": "Chetan",
-            "num_copies": 0
-        },
-        {
-            "bookId": 15,
-            "userName": "Me",
-            "num_copies": 0,
-            "userId": 4
-        },
-        {
-            "bookId": 1,
-            "userName": "Pradeep",
-            "num_copies": 1,
-            "userId": 5
-        },
-        {
-            "bookId": 1,
-            "userId": 101,
-            "num_copies": 1,
-            "allocate_Date": "2025-02-18",
-            "return_Date": "2025-03-01"
-        },
-        {
-            "bookId": 1,
-            "userId": 101,
-            "num_copies": 1,
-            "allocate_Date": "2025-02-18",
-            "return_Date": "2025-03-01"
-        },
-        {
-            "bookId": 1,
-            "userId": 101,
-            "num_copies": 1,
-            "allocate_Date": "2025-02-18",
-            "return_Date": "2025-03-01"
-        },
-        {
-            "bookId": 15,
-            "userId": 1,
-            "num_copies": 0,
-            "issue_Date": "2025-02-18",
-            "return_Date": "2025-02-19"
-        }
-    ]
+  "allocation": [
+    {
+      "bookId": 1,
+      "userId": 101,
+      "num_copies": 1,
+      "allocate_Date": "2025-02-18",
+      "return_Date": "2025-03-01"
+    },
+    {
+      "bookId": 1,
+      "userId": 101,
+      "num_copies": 1,
+      "allocate_Date": "2025-02-18",
+      "return_Date": "2025-03-01"
+    },
+    {
+      "bookId": 1,
+      "userId": 101,
+      "num_copies": 1,
+      "allocate_Date": "2025-02-18",
+      "return_Date": "2025-03-01"
+    },
+    {
+      "bookId": 15,
+      "userId": 1,
+      "num_copies": 0,
+      "issue_Date": "2025-02-20",
+      "return_Date": "2025-02-21"
+    },
+    {
+      "bookId": 15,
+      "userId": 101,
+      "num_copies": 1,
+      "issue_Date": "2025-02-20",
+      "return_Date": "2025-02-20"
+    },
+    {
+      "bookId": 15,
+      "userId": 102,
+      "num_copies": 1,
+      "issue_Date": "2025-02-20",
+      "return_Date": "2025-02-20"
+    },
+    {
+      "bookId": 15,
+      "userId": 1,
+      "num_copies": 1,
+      "issue_Date": "2025-02-20",
+      "return_Date": "2025-02-20"
+    },
+    {
+      "bookId": 15,
+      "userId": 2,
+      "num_copies": 1,
+      "issue_Date": "2025-02-20",
+      "return_Date": "2025-02-20"
+    }
+  ]
 }
