+++ library/src/AddBook.js
@@ -4,6 +4,7 @@ import { useState } from "react";
 function AddBook() {
 
     const [newBook, setBook] = useState({
+        id:0,
         book_name: "",
         author: "",
         num_copies: 0
